+++ src/FSharpVSPowerTools.Logic.VS2015/PeekDefinition.fs
@@ -10,20 +10,20 @@ open Microsoft.VisualStudio.Utilities
 open Microsoft.VisualStudio.Shell.Interop
 open FSharpVSPowerTools
 open FSharpVSPowerTools.ProjectSystem
+open FSharpVSPowerTools.Navigation
 open Microsoft.FSharp.Compiler
 open Microsoft.FSharp.Compiler.SourceCodeServices
 open EnvDTE
 open Microsoft.VisualStudio.Text.Editor
 
-type internal DefinitionPeekableItem(span: SnapshotSpan, range: Range.range, peekResultFactory: IPeekResultFactory) =
+type internal DefinitionPeekableItem(peekResultFactory: IPeekResultFactory, span: SnapshotSpan, range: Range.range, isReadOnly) =
     interface IPeekableItem with
         member __.DisplayName = null
         member __.Relationships = seq [PredefinedPeekRelationships.Definitions]
         member __.GetOrCreateResultSource _relationshipName = 
             { new IPeekResultSource with
                 member __.FindResults(relationshipName, resultCollection, _ct, _callback) =
-                    if relationshipName <> PredefinedPeekRelationships.Definitions.Name then ()
-                    else 
+                    if relationshipName = PredefinedPeekRelationships.Definitions.Name then
                         let filePath = range.FileName
                         let fileName = Path.GetFileName filePath
                         let label = sprintf "%s - (%d, %d)" filePath range.StartLine range.StartColumn
@@ -41,22 +41,22 @@ type internal DefinitionPeekableItem(span: SnapshotSpan, range: Range.range, pee
                                                 range.EndColumn,
                                                 span.StartLine.LineNumber,
                                                 span.StartColumn,
-                                                false)
+                                                isReadOnly)
 
                         resultCollection.Add result
-    //                callback.ReportProgress(100 * ++processedSourceLocations / sourceLocations.Count);
             }
 
 type PeekableItemSource
     (
-        buffer: ITextBuffer,
+        textBuffer: ITextBuffer,
         doc: ITextDocument,
         peekResultFactory: IPeekResultFactory,
+        metadataService: NavigateToMetadataService,
         projectFactory: ProjectFactory,
         vsLanguageService: VSLanguageService
     ) =
 
-    let project = lazy (projectFactory.CreateForDocument buffer doc.FilePath)
+    let project = lazy (projectFactory.CreateForDocument textBuffer doc.FilePath)
 
     let getDefinitionRange (point: SnapshotPoint) =
         async {
@@ -71,14 +71,16 @@ type PeekableItemSource
             | Some (project, span, symbol) ->
                 let! symbolUse = vsLanguageService.GetFSharpSymbolUse(span, symbol, doc.FilePath, project, AllowStaleResults.MatchingSource)
                 match symbolUse with
-                | Some (_fsSymbolUse, fileScopedCheckResults) ->
+                | Some (fsSymbolUse, fileScopedCheckResults) ->
                     let start = span.Start
                     let lineStr = start.GetContainingLine().GetText()
                     let! findDeclResult = fileScopedCheckResults.GetDeclarationLocation(symbol.Line, symbol.RightColumn, lineStr, symbol.Text, preferSignature=false)
-                    return
-                        match findDeclResult with
-                        | FSharpFindDeclResult.DeclFound range -> Some (span, range)
-                        | _ -> None
+                    match findDeclResult with
+                    | FSharpFindDeclResult.DeclFound range -> 
+                        return Some (span, range, false)
+                    | FSharpFindDeclResult.DeclNotFound _ ->
+                        let! range = metadataService.TryFindMetadataRange(project, textBuffer, fileScopedCheckResults.ParseTree, span, fsSymbolUse)
+                        return range |> Option.map (fun r -> span, r, true)
                 | _ -> return None
             | _ -> return None
         }
@@ -88,9 +90,9 @@ type PeekableItemSource
             asyncMaybe {
                 do! if String.Equals(session.RelationshipName, PredefinedPeekRelationships.Definitions.Name, 
                                      StringComparison.OrdinalIgnoreCase) then Some() else None
-                let! triggerPoint = session.GetTriggerPoint(buffer.CurrentSnapshot) |> Option.ofNullable
-                let! symbolSpan, definitionRange = getDefinitionRange triggerPoint
-                peekableItems.Add (DefinitionPeekableItem(symbolSpan, definitionRange, peekResultFactory))
+                let! triggerPoint = session.GetTriggerPoint(textBuffer.CurrentSnapshot) |> Option.ofNullable
+                let! symbolSpan, definitionRange, isReadOnly = getDefinitionRange triggerPoint
+                peekableItems.Add (DefinitionPeekableItem(peekResultFactory, symbolSpan, definitionRange, isReadOnly))
             } |> Async.RunSynchronously |> ignore
 
         member __.Dispose() = ()
@@ -105,6 +107,7 @@ type PeekableItemSourceProvider
     (peekResultFactory: IPeekResultFactory,
      textDocumentFactoryService: ITextDocumentFactoryService,
      [<Import(typeof<SVsServiceProvider>)>] serviceProvider: System.IServiceProvider,
+     metadataService: NavigateToMetadataService,
      projectFactory: ProjectFactory,
      vsLanguageService: VSLanguageService) =
 
@@ -125,5 +128,5 @@ type PeekableItemSourceProvider
                     match textDocumentFactoryService.TryGetTextDocument buffer with
                     | true, doc ->
                         buffer.Properties.GetOrCreateSingletonProperty(
-                            fun () -> upcast new PeekableItemSource(buffer, doc, peekResultFactory, projectFactory, vsLanguageService))
+                            fun () -> upcast new PeekableItemSource(buffer, doc, peekResultFactory, metadataService, projectFactory, vsLanguageService))
                     | _ -> null