+++ src/FSharpVSPowerTools.Logic/GoToDefinitionFilter.fs
@@ -37,25 +37,27 @@ type GoToDefinitionFilter
     (
         textDocument: ITextDocument,
         view: IWpfTextView, 
-        editorOptionsFactory: IEditorOptionsFactoryService, 
         vsLanguageService: VSLanguageService, 
         serviceProvider: System.IServiceProvider,                          
         projectFactory: ProjectFactory,
         referenceSourceProvider: ReferenceSourceProvider,
+        metadataService: NavigateToMetadataService,
         navigationPreference: NavigationPreference,
         fireNavigationEvent: bool
     ) =
     
     let urlChanged = if fireNavigationEvent then Some (Event<UrlChangeEventArgs>()) else None
     let mutable currentUrl = None
+
     let dte = serviceProvider.GetDte()
-    let project = lazy (projectFactory.CreateForDocument view.TextBuffer textDocument.FilePath)
+    let textBuffer = view.TextBuffer
+    let project = lazy (projectFactory.CreateForDocument textBuffer textDocument.FilePath)
 
     let getDocumentState () =
         async {
             let projectItems = maybe {
                 let! project = project.Value
-                let! caretPos = view.TextBuffer.GetSnapshotPoint view.Caret.Position
+                let! caretPos = textBuffer.GetSnapshotPoint view.Caret.Position
                 let! span, symbol = vsLanguageService.GetSymbol(caretPos, textDocument.FilePath, project)
                 return project, span, symbol }
 
@@ -67,235 +69,8 @@ type GoToDefinitionFilter
                     let lineStr = span.Start.GetContainingLine().GetText()
                     let! findDeclResult = fileScopedCheckResults.GetDeclarationLocation(symbol.Line, symbol.RightColumn, lineStr, symbol.Text, preferSignature=false)
                     return Some (project, fileScopedCheckResults.ParseTree, span, fsSymbolUse, findDeclResult) 
-                | _ -> return None
-            | _ -> return None
-        }
-
-    // Use a single cache across text views
-    static let xmlDocCache = Dictionary<string, IVsXMLMemberIndex>()
-    let xmlIndexService = serviceProvider.GetService<IVsXMLMemberIndexService, SVsXMLMemberIndexService>()
-
-    /// If the XML comment starts with '<' not counting whitespace then treat it as a literal XML comment.
-    /// Otherwise, escape it and surround it with <summary></summary>
-    let processXml (xml: string) =
-        if String.IsNullOrEmpty(xml) then xml
-        else
-            let trimmedXml = xml.TrimStart([|' ';'\r';'\n'|])
-            if String.IsNullOrEmpty(trimmedXml) then xml
-            else
-                match trimmedXml.[0] with
-                | '<' ->
-                    String.Join("", "<root>", xml, "</root>")
-                | _ ->
-                    let escapedXml = SecurityElement.Escape(xml)
-                    String.Join("", "<summary>", escapedXml, "</summary>")
-            
-    let getXmlDocBySignature (fsSymbol: FSharpSymbol) signature =
-        match fsSymbol.Assembly.FileName with
-        | Some assemblyName ->
-            match xmlDocCache.TryGetValue(assemblyName) with
-            | true, xmlIndex -> 
-                match xmlIndex.ParseMemberSignature(signature) with
-                | _, 0u -> []
-                | _, index ->
-                    match xmlIndex.GetMemberXML(index) with
-                    | VSConstants.S_OK, xml ->
-                        let processedXml = processXml xml
-                        let xmlDocs = ResizeArray()
-                        match xmlIndexService.GetMemberDataFromXML(processedXml) with
-                        | VSConstants.S_OK, memberData ->
-                            match memberData.GetSummaryText() with
-                            | VSConstants.S_OK, xmlSummary ->
-                                xmlDocs.Add(xmlSummary)
-                            | _ -> ()
-                            match memberData.GetParamCount() with
-                            | VSConstants.S_OK, count when count > 0 ->
-                                for i in 0..count-1 do
-                                    match memberData.GetParamTextAt(i) with
-                                    | VSConstants.S_OK, name, text ->
-                                        let xmlDoc = sprintf "%s: %s" name text
-                                        xmlDocs.Add(xmlDoc)
-                                    | _ -> ()
-                            | _ -> ()
-                            xmlDocs |> Seq.toList
-                        | _ -> []                            
-                    | _ -> []
-            | false, _ -> 
-                match xmlIndexService.CreateXMLMemberIndex(assemblyName) with
-                | VSConstants.S_OK, xmlIndex ->
-                    match xmlIndex.BuildMemberIndex() with
-                    | VSConstants.S_OK ->
-                        xmlDocCache.Add(assemblyName, xmlIndex)
-                        []
-                    | _ -> []
-                | _ -> []
-        | None -> []
-
-    // Keep a single window frame for all text views
-    static let mutable currentWindow: (IVsWindowFrame * string) option = None 
-
-    let gotoExactLocation signature filePath signatureProject currentSymbol vsTextBuffer =
-        async {
-            let! symbolUses = 
-                vsLanguageService.GetAllUsesOfAllSymbolsInSourceString(signature, filePath, signatureProject, AllowStaleResults.No, false)
-
-            /// Try to reconstruct fully qualified name for the purpose of matching symbols
-            let rec tryGetFullyQualifiedName (symbol: FSharpSymbol) = 
-                Option.attempt (fun _ -> 
-                    match symbol with
-                    | TypedAstPatterns.Entity (entity, _, _) ->
-                        Some (sprintf "%s.%s" entity.AccessPath entity.DisplayName)
-                    | MemberFunctionOrValue mem ->
-                        tryGetFullyQualifiedName mem.EnclosingEntity
-                        |> Option.map (fun parent -> sprintf "%s.%s" parent mem.DisplayName)
-                    | Field(field, _) ->
-                        tryGetFullyQualifiedName field.DeclaringEntity
-                        |> Option.map (fun parent -> sprintf "%s.%s" parent field.DisplayName)
-                    | UnionCase uc ->
-                        match uc.ReturnType with
-                        | TypeWithDefinition entity ->
-                            tryGetFullyQualifiedName entity
-                            |> Option.map (fun parent -> sprintf "%s.%s" parent uc.DisplayName)
-                        | _ -> 
-                            None
-                    | ActivePatternCase case ->
-                        let group = case.Group
-                        group.EnclosingEntity
-                        |> Option.bind tryGetFullyQualifiedName
-                        |> Option.map (fun parent -> 
-                            let sb = StringBuilder()
-                            sb.Append("|") |> ignore
-                            for name in group.Names do
-                                sb.AppendFormat("{0}|", name) |> ignore
-                            if not group.IsTotal then
-                                sb.Append("_|") |> ignore
-                            sprintf "%s.( %O )" parent sb)
-                    | _ ->
-                        None)
-                |> Option.flatten
-                |> Option.orTry (fun _ -> Option.attempt (fun _ -> symbol.FullName))
-
-            let isLocalSymbol filePath (symbol: FSharpSymbol) =
-                symbol.DeclarationLocation 
-                |> Option.map (fun r -> String.Equals(r.FileName, filePath, StringComparison.OrdinalIgnoreCase)) 
-                |> Option.getOrElse false
-
-            let currentSymbolFullName = tryGetFullyQualifiedName currentSymbol
-
-            let matchedSymbol = 
-                symbolUses 
-                |> Seq.groupBy (fun { SymbolUse = symbolUse } -> symbolUse.Symbol)
-                |> Seq.collect (fun (_, uses) -> Seq.truncate 1 uses)
-                |> Seq.choose (fun { SymbolUse = symbolUse } -> 
-                    match symbolUse.Symbol with
-                    | TypedAstPatterns.Entity _
-                    | MemberFunctionOrValue _
-                    | ActivePatternCase _                   
-                    | UnionCase _
-                    | Field _ as symbol -> 
-                        let symbolFullName = tryGetFullyQualifiedName symbol
-                        if symbolFullName = currentSymbolFullName && isLocalSymbol filePath symbol then
-                            Some symbol
-                        else None
-                    | _ -> None)
-                |> Seq.sortBy (fun symbol -> 
-                    symbol.DeclarationLocation |> Option.map (fun r -> r.StartLine, r.StartColumn))
-                |> Seq.tryHead
-
-            match matchedSymbol with
-            | Some symbol ->
-                let vsTextManager = serviceProvider.GetService<IVsTextManager, SVsTextManager>()
-                symbol.DeclarationLocation
-                |> Option.iter (fun r -> 
-                    let (startRow, startCol) = (r.StartLine-1, r.StartColumn)
-                    vsTextManager.NavigateToLineAndColumn(vsTextBuffer, ref Constants.guidLogicalTextView, startRow, startCol, startRow, startCol) 
-                    |> ensureSucceeded)
-            | None ->
-                Logging.logInfo (fun _ -> sprintf "Can't find a matching symbol for '%A'" currentSymbol)
-        }
-
-    // Now the input is an entity or a member/value.
-    // We always generate the full enclosing entity signature if the symbol is a member/value
-    let navigateToMetadata project (span: SnapshotSpan) ast (fsSymbolUse: FSharpSymbolUse) = 
-        async {
-            let fsSymbol = fsSymbolUse.Symbol
-            let displayContext = fsSymbolUse.DisplayContext
-            let fileName = SignatureGenerator.getFileNameFromSymbol fsSymbol
-
-            // The file system is case-insensitive so list.fsi and List.fsi can clash
-            // Thus, we generate a tmp subfolder based on the hash of the filename
-            let subFolder = string (uint32 (hash fileName))
-
-            let filePath = Path.Combine(Path.GetTempPath(), subFolder, fileName)
-            let statusBar = serviceProvider.GetService<IVsStatusbar, SVsStatusbar>()
-            let editorOptions = editorOptionsFactory.GetOptions(view.TextBuffer)
-            let indentSize = editorOptions.GetOptionValue((IndentSize()).Key)  
-            match VsShellUtilities.IsDocumentOpen(serviceProvider, filePath, Constants.guidLogicalTextView) with
-            | true, _hierarchy, _itemId, windowFrame ->
-                let vsTextView = VsShellUtilities.GetTextView(windowFrame)
-                let mutable vsTextLines = Unchecked.defaultof<_>
-                vsTextView.GetBuffer(&vsTextLines) |> ensureSucceeded
-                let vsTextBuffer = vsTextLines :> IVsTextBuffer                
-                match currentWindow, projectFactory.TryGetSignatureProjectProvider(filePath) with
-                | Some (_, signature), Some signatureProject ->
-                    do! gotoExactLocation signature filePath signatureProject fsSymbol vsTextBuffer
-                | _ -> 
-                    Logging.logInfo (fun _ -> sprintf "Can't find a signature or signature project for '%s'" filePath)
-                // If the buffer has been opened, we will not re-generate signatures
-                windowFrame.Show() |> ensureSucceeded
-            | _ ->
-                let (startLine, startCol, _, _) = span.ToRange()
-                let pos = mkPos (startLine+1) startCol
-                let openDeclarations = 
-                    ast 
-                    |> Option.map (OpenDeclarationGetter.getEffectiveOpenDeclarationsAtLocation pos) 
-                    |> Option.getOrElse []
-                match SignatureGenerator.formatSymbol 
-                        (getXmlDocBySignature fsSymbol) indentSize displayContext openDeclarations fsSymbol 
-                        SignatureGenerator.Filterer.NoFilters SignatureGenerator.BlankLines.Default with
-                | Some signature ->
-                    let directoryPath = Path.GetDirectoryName(filePath)
-                    Directory.CreateDirectory(directoryPath) |> ignore
-                    File.WriteAllText(filePath, signature)
-                    let mutable hierarchy = Unchecked.defaultof<_>
-                    let mutable itemId = Unchecked.defaultof<_>
-                    let windowFrame = ref null
-                    let canShow = 
-                        try
-                            VsShellUtilities.OpenDocument(
-                                serviceProvider, 
-                                filePath, 
-                                Constants.guidLogicalTextView, 
-                                &hierarchy,
-                                &itemId,
-                                windowFrame)
-                            true
-                        with _ -> false
-                    if canShow then
-                        // Ensure that only one signature is opened at a time
-                        currentWindow 
-                        |> Option.bind (fun (window, _) -> Option.ofNull window)
-                        |> Option.iter (fun window -> window.CloseFrame(uint32 __FRAMECLOSE.FRAMECLOSE_NoSave) |> ignore)
-                        // Prevent the window being reopened as a part of a solution
-                        (!windowFrame).SetProperty(int __VSFPROPID5.VSFPROPID_DontAutoOpen, true) |> ignore
-                        currentWindow <- Some (!windowFrame, signature)
-                        let vsTextView = VsShellUtilities.GetTextView(!windowFrame)
-                        let mutable vsTextLines = Unchecked.defaultof<_>
-                        vsTextView.GetBuffer(&vsTextLines) |> ensureSucceeded
-                        let vsTextBuffer = vsTextLines :> IVsTextBuffer
-                        match vsTextBuffer.GetStateFlags() with
-                        | VSConstants.S_OK, currentFlags ->
-                            // Try to set buffer to read-only mode
-                            vsTextBuffer.SetStateFlags(currentFlags ||| uint32 BUFFERSTATEFLAGS.BSF_USER_READONLY) |> ignore
-                        | _ -> ()
-                        let signatureProject = projectFactory.RegisterSignatureProjectProvider(filePath, project)
-                        do! gotoExactLocation signature filePath signatureProject fsSymbol vsTextBuffer
-                        // We display the window after putting the project into the project system.
-                        // Hopefully syntax coloring on generated signatures will catch up on time.
-                        (!windowFrame).Show() |> ensureSucceeded
-                        statusBar.SetText(Resource.goToDefinitionStatusMessage) |> ignore
-                | None ->
-                    statusBar.SetText(Resource.goToDefinitionInvalidSymbolMessage) |> ignore  
+                | None -> return None
+            | None -> return None
         }
 
     let shouldGenerateDefinition (fsSymbol: FSharpSymbol) =
@@ -306,7 +81,7 @@ type GoToDefinitionFilter
         | Entity(Enum as e, _, _) when not e.IsFSharp -> false
         | _ -> true
 
-    let replace (b:string) c (a:string) = a.Replace(b, c)
+    let replace (oldValue: string) newValue (str: string) = str.Replace(oldValue, newValue)
 
     let getSymbolCacheDir() = 
         let keyName = String.Format(@"Software\Microsoft\VisualStudio\{0}.0\Debugger",
@@ -386,7 +161,7 @@ type GoToDefinitionFilter
                 // Fsi files don't appear on symbol servers, so we try to get urls via its associated fs files
                 let path, isFsiFile = 
                     let path = fullPathOf range.FileName
-                    let isFsiFile = path.EndsWith(".fsi", StringComparison.OrdinalIgnoreCase)
+                    let isFsiFile = isSignatureFile path
                     let newPath = if isFsiFile then changeExt ".fs" path else path
                     newPath, isFsiFile
                 let! url = pdbReader.GetDownloadUrl(path)
@@ -422,7 +197,7 @@ type GoToDefinitionFilter
                     match directive with
                     | Some fileToOpen ->
                         // directive found, navigate to the file at first line
-                        serviceProvider.NavigateTo(fileToOpen, 0, 0, 0, 0)
+                        serviceProvider.NavigateTo(fileToOpen, startRow=0, startCol=0, endRow=0, endCol=0)
                     | None ->
                         // Run the operation on UI thread since continueCommandChain may access UI components.
                         do! Async.SwitchToContext uiContext
@@ -432,7 +207,7 @@ type GoToDefinitionFilter
                     match navigationPreference with
                     | NavigationPreference.Metadata ->
                         if shouldGenerateDefinition fsSymbolUse.Symbol then
-                            return! navigateToMetadata project span parseTree fsSymbolUse
+                            return! metadataService.NavigateToMetadata(project, textBuffer, parseTree, span, fsSymbolUse)
                     | NavigationPreference.SymbolSourceOrMetadata
                     | NavigationPreference.SymbolSource as pref ->   
                         let symbol = fsSymbolUse.Symbol
@@ -456,16 +231,13 @@ type GoToDefinitionFilter
                                 match pref with
                                 | NavigationPreference.SymbolSourceOrMetadata ->
                                     if shouldGenerateDefinition symbol then
-                                        return! navigateToMetadata project span parseTree fsSymbolUse
+                                        return! metadataService.NavigateToMetadata(project, textBuffer, parseTree, span, fsSymbolUse)
                                 | _ ->
                                     let statusBar = serviceProvider.GetService<IVsStatusbar, SVsStatusbar>()
                                     statusBar.SetText(Resource.goToDefinitionNoSourceSymbolMessage) |> ignore
                                     return ()
             }
         Async.StartInThreadPoolSafe (worker, cancelToken.Token)
-
-    static member ClearXmlDocCache() =
-        xmlDocCache.Clear()
 
     member internal __.UrlChanged = urlChanged |> Option.map (fun event -> event.Publish)
     member internal __.CurrentUrl = currentUrl