+++ src/FSharpVSPowerTools/Commands/GotoDefinitionFilterProvider.cs
@@ -38,6 +38,7 @@ namespace FSharpVSPowerTools
         private readonly IEditorOptionsFactoryService _editorOptionsFactory;
         private readonly ProjectFactory _projectFactory;
         private readonly ReferenceSourceProvider _referenceSourceProvider;
+        private readonly NavigateToMetadataService _navigationService;
         private readonly System.IServiceProvider _serviceProvider;
         private readonly SolutionEvents _solutionEvents;
 
@@ -50,6 +51,7 @@ namespace FSharpVSPowerTools
             IEditorOptionsFactoryService editorOptionsFactory,
             ITextDocumentFactoryService textDocumentFactoryService,
             [Import(typeof(DotNetReferenceSourceProvider))] ReferenceSourceProvider referenceSourceProvider,
+            NavigateToMetadataService metadataService,
             VSLanguageService languageService,
             ProjectFactory projectFactory)
         {
@@ -58,6 +60,7 @@ namespace FSharpVSPowerTools
             _editorOptionsFactory = editorOptionsFactory;
             _textDocumentFactoryService = textDocumentFactoryService;
             _referenceSourceProvider = referenceSourceProvider;
+            _navigationService = metadataService;
             _fsharpVsLanguageService = languageService;
             _projectFactory = projectFactory;
 
@@ -72,7 +75,7 @@ namespace FSharpVSPowerTools
 
         private void Cleanup()
         {
- 	        GoToDefinitionFilter.ClearXmlDocCache();
+            NavigateToMetadataService.ClearXmlDocCache();
         }
 
         public void VsTextViewCreated(IVsTextView textViewAdapter)
@@ -85,8 +88,7 @@ namespace FSharpVSPowerTools
         internal GoToDefinitionFilter RegisterCommandFilter(IWpfTextView textView, bool fireNavigationEvent)
         {
             var textViewAdapter = _editorFactory.GetViewAdapter(textView);
-            if (textViewAdapter == null) return null;
-            return Register(textViewAdapter, textView, fireNavigationEvent);
+            return textViewAdapter == null ? null : Register(textViewAdapter, textView, fireNavigationEvent);
         }
 
         private GoToDefinitionFilter Register(IVsTextView textViewAdapter, IWpfTextView textView, bool fireNavigationEvent)
@@ -100,9 +102,9 @@ namespace FSharpVSPowerTools
             ITextDocument doc;
             if (_textDocumentFactoryService.TryGetTextDocument(textView.TextBuffer, out doc))
             {
-                var commandFilter = new GoToDefinitionFilter(doc, textView, _editorOptionsFactory,
-                                                             _fsharpVsLanguageService, _serviceProvider, _projectFactory,
-                                                             _referenceSourceProvider, preference, fireNavigationEvent);
+                var commandFilter = new GoToDefinitionFilter(doc, textView, _fsharpVsLanguageService,
+                                                             _serviceProvider, _projectFactory, _referenceSourceProvider,
+                                                             _navigationService, preference, fireNavigationEvent);
                 if (!_referenceSourceProvider.IsActivated && generalOptions.GoToSymbolSourceEnabled)
                     _referenceSourceProvider.Activate();
                 textView.Properties.AddProperty(serviceType, commandFilter);