+++ RELEASE_NOTES.md
@@ -1,7 +1,8 @@
 #### 2.4.0 - Unreleased
 * Enable peek definition for standalone F# files [#1329](https://github.com/fsprojects/VisualFSharpPowerTools/pull/1329)
 * Optimize on caret move [#1349](https://github.com/fsprojects/VisualFSharpPowerTools/pull/1349)
-* Remove Visual dependencies on F# Tools [#1355](https://github.com/fsprojects/VisualFSharpPowerTools/pull/1355)
+* Remove dependencies on Visual F# Tools [#1355](https://github.com/fsprojects/VisualFSharpPowerTools/pull/1355)
+* View metadata when Peek Definitions on external assemblies [#1369](https://github.com/fsprojects/VisualFSharpPowerTools/pull/1369)
 
 #### 2.3.0 - January 4 2016
 * Implement Peek definition [#1286](https://github.com/fsprojects/VisualFSharpPowerTools/pull/1286)